+++ ChangeLog.md
@@ -13,4 +13,5 @@
   typeclasses can't be implemented with Either or require arbitrary tradeoffs
 
 - 2020-03-04: Reps 5 min
+- 2020-03-09: Reps 10 min - stuck on dumb issue with Semigroup